requirements.txt
python-telegram-bot
langchain
langchain-core
langchain-community
langchain-google-genai
langgraph
langgraph-checkpoint-redis
redis
psycopg2-binary
requests
pandas
ta
python-dotenv
pydantic
scikit-learn
xgboost
APScheduler
matplotlib
numpy
